Initialize static processorInfo once
Currently we are repeatedly initializing the static processorInfo
in every CodeGenerator object. This may cause a race condition.
Note that this is the old version of processor detection on x86
and it is getting removed very soon. The reason this change is needed is
because currently there are assert tests in place that compares new version
cpu detection APIs against old cpu detection APIs which means we still need
to make sure old APIs are working properly until it gets removed.
Gives us the option to disable these old vs. new detection assert tests via an environment
variable "TR_DisableOldVersionCPUDetectionTest"

Signed-off-by: Harry Yu <harryyu1994@gmail.com>